MQ OTR now as a seperate package
- make way for new features
- change soh-otr to soh-otr-mq
- small fix extract dir
- add python make depends

+++ PKGBUILD
@@ -3,15 +3,16 @@
 _romname=gc_mq_d
 _romdesc="Ocarina of Time Master Quest Debug"
 pkgname=soh-otr-${_romname}
-pkgver=1
+pkgver=2
 pkgrel=1
 pkgdesc="OTR game data for SoH (${_romdesc})"
 arch=("any")
 url="https://github.com/HarbourMasters/Shipwright"
-provides=("soh-otr")
-conflicts=("soh-otr")
+license=("unknown")
+provides=("soh-otr-mq")
+conflicts=("soh-otr-mq")
 depends=("soh")
-makedepends=("soh-otr-exporter")
+makedepends=("soh-otr-exporter" "python")
 source=("verify_baserom.py")
 sha256sums=('bd82ccb54e94505549af7dc07b56dc83b3e63e1c65d525b6fb98ed1b56b8aecf')
 
@@ -34,18 +35,19 @@ prepare() {
 build() {
   cd "${srcdir}"
 
-  ln -s ${SHIP_PREFIX}/assets assets
-  mkdir -p Extract/assets
-  cp -r "${SHIP_PREFIX}"/assets/game/* Extract/assets/
+  [ -e oot.otr ] && rm oot.otr
+  [ ! -e assets ] && ln -s ${SHIP_PREFIX}/assets assets
+  [ -e Extract ] && rm -r Extract; mkdir Extract
+  cp -r "${SHIP_PREFIX}"/assets/game Extract/assets
 
   "assets/extractor/ZAPD.out" ed -eh -i "assets/extractor/xmls/${_romname^^}" \
-    -b baserom.z64 -fl "assets/extractor/filelists" -o placeholder -osf placeholder -gsf 1 \
-    -rconf "assets/extractor/Config_${_romname^^}.xml" -se OTR
+    -b baserom.z64 -fl "assets/extractor/filelists" -o . -osf . -gsf 1 \
+    -rconf "assets/extractor/Config_${_romname^^}.xml" -se OTR --otrfile oot-mq.otr
 }
 
 package() {
   cd "${srcdir}"
 
-  install -Dm644 oot.otr -t "${pkgdir}/${SHIP_PREFIX}"
+  install -Dm644 oot-mq.otr -t "${pkgdir}/${SHIP_PREFIX}"
 }
